model/views: Register handler for subscription events for (un)muting.
We want to be notified about changes in subscriptions like
muting/unmuting of a stream, hence we register the event_type
'subscription' to the event queue on startup, with a handler method
which responds to these events appropriately.

We store a mapping between stream_ids and buttons so that retrieval of
buttons is easier when a change in subscription event occurs.

Old tests for 'event_types' amended.

New tests for handler added.