README/keys: Categorise keys according to usage.
Keys are segregated into the following categories:
- General
- Navigation
- Searching
- Message Actions
- Composing a Message

This involves splitting the table in the README, and tagging each entry
in keys.py with a category.

A dict provides a longer textual description of each category.

Some entries are also adjusted slightly for capitalization.

Adjusted to remove duplicate entries by neiljp.