github: Add workflow cancelling earlier workflows on the same branch.
This builds upon the cancel-previous-runs.yml workflow file from
zulip/zulip, simplifying it for now to hardcode the workflows to cancel,
and adapting it to zulip/zulip-terminal.

As with the zulip/zulip workflow, it does not run for pushes to the
original repository, to ensure CI completes on every push to at least
the main branch.

While simplified, original work was by Priyank Patel <priyankp390@gmail.com>

In future we could extend back to auto-detecting runs from all
workflows, or simplify and use the cancellation in each workflow as
detailed at https://github.com/styfle/cancel-workflow-action#usage