views: Remove dead code in load_old_messages & simplify tests.
This removes a special case where there is no focus and yet we load old
messages. It's unclear where this derives from, there is no similar case
for new messages, and no comment or commit message to explain this.

The simplification makes the new and old message loading cases very
similar.

The tests when users scroll outside of the list were previously multiple
tests; this commit adjust them to be explicitly parametrized.