keys: Add hotkey 'i' for viewing user information popup.
As a v0, user info would be seen by pressing `i` in the users' list.
This would allow the unification of the keypress event for `i` in all
three views.

It could be integrated to also trigger from message info view as a
follow up, potentially with a different hotkey.

docs/hotkeys.md regenerated.