boxes: Create a visual "box" by padding whitespaces.
This commit moves the previous code into `parse_codeblock`, a static
method which creates a visual "box" by:
* Finding the longest line using `code_soup.text`.
* Appending `span` text into markup
* Adding the NavigableStrings not contributing to line
breaks into markup.
* Adding `right_padding` to the end of each line.
* Making sure multiple empty lines are padded using `padded_line`.
* Adding `code_indent` part of the NavigableString to markup at the
start of a new line.

The NavigableString is added in parts and not all at once because
though it would produce the same output, it would later help while
implementing `hl_lines` preventing the case where half the previous
line is highlighted along with the line expected.

Tests amended.