boxes: Restrict padding to avoid wrapping.
This commit sets a limit to how much padding is to be added. Many a
times most of the code has smaller lines and there is one line which
is very long which causes all the lines to wrap and this makes the
"box" look very messy. To avoid this a `MAX_NO_WRAP_WIDTH` constant
of 70 is used which is a reasonable limit for in a full screen
no autohide mode.