bugfix: boxes: Don't send typing updates to oneself.
Prior to this commit, typing updates were sent to all recipients of
the current private message as specified in `recipient_user_ids`,
which is an unnecesary expense for the server though minute.

This commit uses another list, `typing_recipient_user_ids`, that
stores all the `recipient user_ids` except for the user's own, that
can be used to send typing updates. This list is initiated as a part
of WriteBox for persistence to send start and stop updates.

Tests added.