model: Handle subscription events for desktop_notifications property.
This commit adds handling for subscription events related to
desktop_notifications. The model._handle_subscription_events now also
checks for events with change in "desktop_notifications" property and
adds/removes those streams from model.visual_notified_streams variable
depending on the value specified in the event. This allows the model to
stay in sync with the server for changes in visual notified streams.

Tests added.