refactor: model: Migrate to new asynchronous Typing Notifications.
This commit refactors `_handle_typing_event` to use the new asynchronous
approach to typing notifications in the Controller.

As per the description in the commit which introduced it, on a start
event we now:
* update the `sender_name` in the `active_notifications_info` state dict
* run the `show_typing_notification` method
  (only if not running - by checking `is_typing_notification_in_progress`)

On a stop event we reset the state dict, which can cause the
notification method to cease running.

Tests added.