refactor: model: Separate Typing Notification from Model.
This commit refactors `_handle_typing_event` to separate the actual
setting of notification as a method in `Controller`.

This is done using an asynchronous function
`show_typing_notification` and a state dictionary
`active_conversation_info` start and end the notification.

To make sure 2 typing async functions don't run parallelly we use a
boolean `is_typing_notification_in_progress`.

`sender_name` is stored so that this can be expanded for group
typing notifications.

Tests added.