refactor: buttons: Improve stream data handling via typed classes.
Currently stream narrow data is being primarily used for two purposes:
 - Validating the narrow link,
 - Patching the relevant information for successful narrowing.

The stream data under validation has either only stream id or only
stream name, which makes it a better candidate for `total=False`.
Thus, this type of data it is extracted to a separate TypedDict class
`DecodedStreamPartial`.

Once the stream data is successfully validated, we're ready to patch
the relevant stream data. This should be strictly typed since now we
know that the stream data is valid. Thus, this type of data is
extracted to a separate TypedDict class `DecodedStreamAbsolute`.

Tests amended.