refactor: model/helper/views: Remove timezone & last_active from bots.
The timezone and last_active fields are not present for bots, so like with
bot-specific data in the bot version of TidiedUserInfo, only store these fields
in the regular user version of TidiedUserInfo.

The timezone and last_active calculations are moved into the user-specific
conditional in Model.get_user_info and in UserInfoView._fetch_user_data.

Fixtures updated.