model/helper: Remove Optional on TidiedUserInfo bot_type.
Since bot_type is only null/None when is_bot is False, it is guaranteed to have
actual type information (int) in the bot version of TidiedUserInfo, so no
longer needs to be Optional.

Assert added in model to ensure we notice if this is incorrect.

Add bot_type to some bot test cases, since it is now required, and remove test
case where the bot_type was assumed valid when None.