refactor: views: Refine focus of user information popup helper function.
The responsibility for acquiring the user data from the model is moved
from the _fetch_user_data helper into the user information popup
initializer. The helper function is then solely responsible for
translating the TidiedUserInfo into a format suitable for output, so is
renamed to _polish_user_data.

This simplifies the tests, though currently involves type casts due to
not using a factory for the TidiedUserInfo.